% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depseudonymize.R
\name{depseudonymize}
\alias{depseudonymize}
\title{Convert a pseudonym back to its input value.}
\usage{
depseudonymize(
  MAINZELLISTE_BASE_URL = NULL,
  MAINZELLISTE_API_KEY = NULL,
  MAINZELLISTE_FIELDNAME = NULL,
  mainzelliste_fieldvalue,
  from_env = FALSE
)
}
\arguments{
\item{MAINZELLISTE_BASE_URL}{(Optional, String)
The URL to your Mainzelliste API.
E.g. 'https://ml.hospital.de'.}

\item{MAINZELLISTE_API_KEY}{(Optional, String)
The API key which is allowed to access the Mainzelliste.}

\item{MAINZELLISTE_FIELDNAME}{(Optional, String) The name of
the field to use for the Mainzelliste. Specified in the ML-config.}

\item{mainzelliste_fieldvalue}{(String) The actual value to de-pseudonymized.}

\item{from_env}{(Optional, Boolean, Default = `FALSE`) If true, the
connection parameters `MAINZELLISTE_BASE_URL`, `MAINZELLISTE_API_KEY` and
`MAINZELLISTE_FIELDNAME` are read from the environment
and can therefore be left empty when calling this function.}
}
\value{
(vector) All pseudonyms for the input values.
}
\description{
Convert a pseudonym back to its input value.
}
